requirements.txt
# Core Dependencies
streamlit==1.29.0
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
ebooklib==0.18
striprtf==0.0.26
beautifulsoup4==4.12.2
Pillow==10.1.0
pytesseract==0.3.10
PyMuPDF==1.23.8

# NLP & AI
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# LLM Integration
openai==1.6.1
anthropic==0.8.1
tiktoken==0.5.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Caching & Performance
redis==5.0.1
diskcache==5.6.3

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring (Optional)
sentry-sdk==1.39.1

# Production Server
gunicorn==21.2.0
uvicorn==0.25.0

# Background Tasks (Future)
celery==5.3.4
redis==5.0.1

# Additional Utilities
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.2
